Merges nbt-serde into the main crate behind a feature flag.
This commit also migrates and/or merges all tests and benchmarks to the
new crate layout.
