+++ src/lib.rs
@@ -9,9 +9,15 @@ pub use error::{Error, Result};
 pub use value::Value;
 
 pub mod raw;
-
 mod blob;
 mod error;
 mod value;
+
+#[cfg(feature = "serde")] #[macro_use] extern crate serde;
+
+#[cfg(feature = "serde")] #[macro_use] mod macros;
+#[cfg(feature = "serde")] pub mod de;
+#[cfg(feature = "serde")] pub mod ser;
+#[cfg(feature = "serde")] pub mod serde_error;
 
 #[cfg(test)] mod tests;
