+++ src/serde/mod.rs
@@ -1,7 +1,3 @@
-#[macro_use] extern crate serde;
-extern crate flate2;
-extern crate nbt;
-
 #[macro_use] mod macros;
 
 pub use error::{Error, Result};
